Project data files
ATLASLST.DB is the program-generated database file which derives its contents from the 4 Db* files. The reading of the entries and the management of the database is done within the application. If you wish to test the capability of reading files and scanning its contents, you may download the Db files and place them in the same directory as the CPP file and specify the path when prompted, while running the application. 

NOTE1: Db1.txt is a huge file (about 1MB in size), and it will take some time to read all its contents. If you don't wish to wait through the process' completion, you may just download and plug in the ATLASLST.DB file,

NOTE2: The program will ask you if the region names are stored in delimiters. If you're adding from Db1.txt, make sure you set the delimiter character to " (double quote). The delimiter here essentially means that the region string of interest is stored between two successive occurences of the specified character. Since Db1.txt has its region entries enclosed in double-quotes, the delimiter is ".